requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
python-dotenv>=1.0.0

# Video extraction
yt-dlp>=2023.1.6
requests>=2.31.0

# AI/OpenAI
openai>=1.0.0

# Template and Configuration Management
jinja2>=3.1.0
pyyaml>=6.0

# Caching (optional)
redis[hiredis]>=4.5.0

# Development/Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Removed unused dependencies
# TikTokApi==7.1.0  # Not used in current implementation
# playwright==1.40.0  # Not used in current implementation